Ability to modify smartly the jsonpatch based on the main json
One of the limitations of jsonpatch is that, for example, if we add
a new value, the path of subkeys where this value live must exist.
If these subkeys don't exist, then the add operation fail.

This commit adds a step to parse a raw patch and wrappers over each
patch operation. The important feature is that, when we find
non-existing subkeys in the add operation, then we push these keys
into the value that we are adding, so they get created.