Add circleci pipeline to publish the docker image to docker hub
The TM org has a circleci context called "conntest-dockerhub-release"
which is already configured to with the necessary credentials to push
images to TM's docker hub.

This commit puts on this new pipeline the responsibility to build and
push new docker images to docker hub when we create a new release. The
old github actions job that did the same task has been deleted.