Refactor the config_parser into multiple composable components
The config_parser.py was a file that contained all the classes and
methods needed to parse the yaml config used to configure the webhook.
Its purpose was to read the yaml file and generate objects that the core
of the app can manage and understand.

The problem with the old approach is that it was difficult to support
multiple formats of the yaml config file. Ideally, if we decide to
modify the format of the yaml config file, the app should be able to
handle both old and new version (backward compatibility is always
welcome for end users).

This commit relocates the classes defined in config_parser.py into
different files within the config_parser directory. Apart from that,
now each subpiece of the yaml file is parsed by an interface. This interface
can have different implementations in case we modify the format of the
underlying yaml. Then, the GenericWebhookConfigManifest class has a
method to instantiate the different subparsers according to the schema
version of the yaml file.