Add Makefile
The logic of some of the scripts under "./scripts/" has been moved to
the new Makefile. This improves the developer experience since this
Makefile provides all the commands a developer needs to write code in
this repo.

Apart from that, this commit also fixes the last lint issues, so now the
linter returns a 10/10 score.