Refactor the operators_test and report test coverage
The operators_test was a badly designed test suite, since it said it was
just testing the operators but in fact it was also testing the process
of parsing them, since it's much easier to specify operations using the
GenericWebhookConfig CRD.

The problem was that the CRD will evolve and the generic-k8s-webhook
should support (and test) both old and new versions of the CRD. The
refactor of the tests makes easier to keep test coverage for all the
versions that are supported.

Apart from that, we're introducing pytest-cov to write a test coverage
report after executing all the unittests.