Create alias for certain operations
This commit improves the name of certain operations. This change has
effect on v1beta1. However, to be able to consume configs written with
v1alph1, we're keeping the old names. The operations that now have a new
alias are:

- "and" -> "all"
- "or" -> "any"
- "forEach" -> "map"

Appart from that, we're changing the behaviour of "or" (and "any) when
they have no operands. Before, the result was true, but now it's false.
This makes this operation consistent with `any([])` in Python. The
intention is to make it's output more intuitive for Python users.