Ability to parse expressions that can filter and/or map a list
These expressions have the following format:

```
<reference> (("|" "->") <expr>)
```

<reference> is a reference to a list in the yaml that we're processing.
What we have on the right hand side is a filter ("|") or a map ("->").
The <expr> is the function used to filter or map the list in the left
hand side. We can have multiple filters and/or maps on the right. For
example, we can see if all the side containers request less than 1 cpu
using the following expression:

```
all: .spec.containers | .name != "main" -> .requests.cpu < 1
```