Refactor the jsonpatch tests
The jsonpatch tests now follow a similar format as the conditions tests,
where we define them in a declarative way in a yaml format. Moreover, we
define tests for specific schema versions. This allows us to keep
running old tests for old schemas and for schemas with backward
compatibility and introduce new tests that are only executed for the new
schemas.