Ability to use expressions in a json patch
Introducing the JsonPatchParserV2. This new json patch parser is
backwards compatible with the already supported json patch operations.
On top of that, it adds the ability to use expressions to generate new
values on a json patch operation. For example, we can add a prefix to
the pod name:

```
op: expr
path: .metadata.name
value: '"prefix-" ++ .metadata.name'
```