Fix problem when adding an element on the index 0 of a list
This problem happened when the list doesn't exist on the original yaml.
In this situation, the new element was added as a dictionary of key "0".
Now, with the current fix, the list is properly created and the element
is added as a normal list element.

Apart from that, we've commented the step that pushes the self-contained
package to the release since we need to enable first
"softprops/action-gh-release@v2".