Ability to set multiple webhooks on the same path
If we define different webhooks with the same value on the "path" field,
then these webhooks will be called in sequence when the server receives
a request for the given path. The resulting "accept" responses will be
ANDed and the patches will be concatenated. Notice that a given webhook
will see the patches from the previous webhooks already applied to the
object that it must inspect.